+++ FE/src/components/SettingContent/CategorySetting/categorySetting.scss
@@ -57,68 +57,4 @@
   .spending__category__setting {
     @include categorySettingBox;
   }
-
-  .category__modal__overlay {
-    position: absolute;
-    width: 100%;
-    height: 100%;
-    transform: translateY(-20%);
-    background-color: rgba(0, 0, 0, 0.9);
-    z-index: 2;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    .category__modal {
-      width: 30%;
-      height: 30%;
-      border: 1px solid #979797;
-      border-radius: 6px;
-      display: flex;
-      flex-direction: column;
-
-      align-items: center;
-      position: relative;
-
-      .modal__title {
-        padding: 1.5em;
-        font-size: 1.5em;
-        margin-bottom: 1.5em;
-        text-shadow: $selectedTextShadow;
-      }
-      .modal__contents {
-        width: 100%;
-        height: 100%;
-        position: absolute;
-        display: flex;
-        justify-content: center;
-        align-items: center;
-        gap: 7%;
-
-        .create__category__input {
-          width: 50%;
-          @include input-form;
-        }
-
-        .edit__category__input {
-          width: 50%;
-          @include input-form;
-        }
-
-        .category__edit__btns {
-          display: flex;
-          flex-direction: column;
-          justify-content: center;
-          height: 100%;
-          gap: 5%;
-          .category__update__btn,
-          .category__delete__btn {
-            @include general-button-design;
-          }
-        }
-        .category__create__btn {
-          @include general-button-design;
-        }
-      }
-    }
-  }
 }
