+++ FE/src/components/SettingContent/CategorySetting/index.tsx
@@ -1,11 +1,9 @@
-import React, { useState, useRef, useEffect } from 'react';
+import React, { useState } from 'react';
 
 import { useTransactionData } from '../../../store/AccountBook/accountBookInfoHook';
-import {
-  createCategory,
-  updateCategory,
-  deleteCategory,
-} from '../../../api/category';
+
+import CategoryModal from './CategoryModal';
+
 import './categorySetting.scss';
 
 export default function CategorySetting({ accountBookId }) {
@@ -20,26 +18,6 @@ export default function CategorySetting({ accountBookId }) {
   const spendingCategories = useTransactionData(store =>
     store.getSpendingCategories(),
   );
-  const changeCategories = useTransactionData(store => store.createCategories);
-  const updateTarget = useTransactionData(store => store.updateCategory);
-  const deleteTarget = useTransactionData(store => store.deleteCategory);
-
-  const createRef = useRef();
-  const editRef = useRef();
-
-  const onClickOverlay = () => {
-    setModal('');
-  };
-
-  const modalSetting = () => {
-    if (modal === 'create') {
-      createRef.current.focus();
-    }
-    if (modal === 'update') {
-      editRef.current.value = editContent;
-      editRef.current.focus();
-    }
-  };
 
   const onClickAddBtn = event => {
     setModal(() => 'create');
@@ -52,52 +30,6 @@ export default function CategorySetting({ accountBookId }) {
     setCategoryType(() => event.target.dataset.type);
     setCategoryId(() => event.target.dataset.categoryid);
   };
-
-  const onClickCreateBtn = () => {
-    createCategory({
-      accountBookId,
-      name: createRef.current.value,
-      type: categoryType,
-      icon: 1,
-    });
-
-    changeCategories({
-      name: createRef.current.value,
-      type: categoryType,
-      icon: 1,
-    });
-
-    setModal(() => '');
-  };
-
-  const onClickUpdateBtn = () => {
-    updateCategory({
-      accountBookId,
-      categoryId,
-      name: editRef.current.value,
-      type: categoryType,
-      icon: 1,
-    });
-
-    updateTarget({
-      categoryId,
-      name: editRef.current.value,
-      type: categoryType,
-      icon: 1,
-    });
-
-    setModal(() => '');
-  };
-
-  const onClickDeleteBtn = () => {
-    deleteCategory({ accountBookId, categoryId });
-    deleteTarget({ categoryId });
-    setModal(() => '');
-  };
-
-  useEffect(() => {
-    modalSetting();
-  }, [modal, incomeCategories, spendingCategories]);
 
   return (
     <div className="category__setting__container">
@@ -133,6 +65,7 @@ export default function CategorySetting({ accountBookId }) {
               <div
                 className="category__unit"
                 data-categoryid={item._id}
+                data-type={item.type}
                 onClick={onClickUnit}
               >
                 {item.name}
@@ -149,54 +82,14 @@ export default function CategorySetting({ accountBookId }) {
         )}
       </div>
       {modal && (
-        <div className="category__modal__overlay" onClick={onClickOverlay}>
-          <div className="category__modal" onClick={e => e.stopPropagation()}>
-            {modal === 'create' ? (
-              <>
-                <div className="modal__title">Create</div>
-                <div className="modal__contents">
-                  <input
-                    ref={createRef}
-                    type="text"
-                    className="create__category__input"
-                    placeholder="Enter Category Name"
-                  />
-                  <button
-                    className="category__create__btn"
-                    onClick={onClickCreateBtn}
-                  >
-                    Create
-                  </button>
-                </div>
-              </>
-            ) : (
-              <>
-                <div className="modal__title">Edit</div>
-                <div className="modal__contents">
-                  <input
-                    ref={editRef}
-                    type="text"
-                    className="edit__category__input"
-                  />
-                  <div className="category__edit__btns">
-                    <button
-                      className="category__update__btn"
-                      onClick={onClickUpdateBtn}
-                    >
-                      Update
-                    </button>
-                    <button
-                      className="category__delete__btn"
-                      onClick={onClickDeleteBtn}
-                    >
-                      Delete
-                    </button>
-                  </div>
-                </div>
-              </>
-            )}
-          </div>
-        </div>
+        <CategoryModal
+          modal={modal}
+          setModal={setModal}
+          accountBookId={accountBookId}
+          categoryType={categoryType}
+          categoryId={categoryId}
+          editContent={editContent}
+        />
       )}
     </div>
   );
